% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{alias_to_symbol_SCE}
\alias{alias_to_symbol_SCE}
\title{Convert aliases to official gene symbols in a SingleCellExperiment Object}
\usage{
alias_to_symbol_SCE(sce, organism)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}

\item{organism}{Is sce data from "mouse" or "human"}
}
\value{
SingleCellExperiment Object
}
\description{
\code{alias_to_symbol_SCE}Convert aliases to official gene symbols in a SingleCellExperiment Object. Makes use of `nichenetr::convert_alias_to_symbols`
}
\examples{
\dontrun{
sce = sce \%>\% alias_to_symbol_SCE("human")
}

}
