% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_analysis.R
\name{calculate_LR_pb_prod_matrix}
\alias{calculate_LR_pb_prod_matrix}
\title{calculate_LR_pb_prod_matrix}
\usage{
calculate_LR_pb_prod_matrix(sce, sample_id, celltype_id, group_id, senders_oi, receivers_oi, fraction_cutoff, lr_network, batches)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}

\item{sample_id}{Name of the meta data column that indicates from which sample/patient a cell comes from}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{group_id}{Metadata column indicating group/condition. If not available: NULL.}

\item{senders_oi}{character vector of the sender cell types of interest}

\item{receivers_oi}{character vector of the receiver cell types of interest}

\item{fraction_cutoff}{Cutoff indicating the minimum fraction of cells of a cell type in a specific sample that are necessary to consider the ligand/receptor as expressed for the final prioritization.}

\item{lr_network}{Prior knowledge Ligand-Receptor network (columns: ligand, receptor)}

\item{batches}{NA if no batches should be corrected for. If there should be corrected for batches during DE analysis and pseudobulk expression calculation, this argument should be the name(s) of the columns in the meta data that indicate the batch(s). Should be categorical. Pseudobulk expression values will be corrected for the first element of this vector.}
}
\value{
lr_prod_mat: matrix giving the pseudobulk expression product of each ligand-receptor-sender-receiver pair
}
\description{
\code{calculate_LR_pb_prod_matrix} Calculate a matrix giving the pseudobulk expression product of each ligand-receptor-sender-receiver pair (that is expressed in at least one sample)
}
\examples{
\dontrun{
library(dplyr)
lr_network = readRDS(url("https://zenodo.org/record/3260758/files/lr_network.rds"))
lr_network = lr_network \%>\% dplyr::rename(ligand = from, receptor = to) \%>\% dplyr::distinct(ligand, receptor)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
senders_oi = "CAF"
receivers_oi = "Malignant"
fraction_cutoff = 0.05
batches = NA
lr_prod_mat = calculate_LR_pb_prod_matrix(sce, sample_id, celltype_id, group_id, senders_oi, receivers_oi, fraction_cutoff, lr_network, batches)
}

}
