% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{combine_sender_receiver_de}
\alias{combine_sender_receiver_de}
\title{combine_sender_receiver_de}
\usage{
combine_sender_receiver_de(sender_de, receiver_de, senders_oi, receivers_oi, lr_network)
}
\arguments{
\item{sender_de}{XXX}

\item{receiver_de}{XXX}

\item{senders_oi}{XXX}

\item{receivers_oi}{XXX}

\item{lr_network}{XXXX}
}
\value{
XXXX
}
\description{
\code{combine_sender_receiver_de}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
