% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{combine_sender_receiver_info_ic}
\alias{combine_sender_receiver_info_ic}
\title{combine_sender_receiver_info_ic}
\usage{
combine_sender_receiver_info_ic(sender_info, receiver_info, senders_oi, receivers_oi, lr_network)
}
\arguments{
\item{sender_info}{XXX}

\item{receiver_info}{XXX}

\item{senders_oi}{XXX}

\item{receivers_oi}{XXX}

\item{lr_network}{XXXX}
}
\value{
XXXX
}
\description{
\code{combine_sender_receiver_info_ic}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
