% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compare_normal_emp_pvals}
\alias{compare_normal_emp_pvals}
\title{compare_normal_emp_pvals}
\usage{
compare_normal_emp_pvals(DE_info, DE_info_emp, adj_pval = FALSE)
}
\arguments{
\item{DE_info}{Output of `get_DE_info`}

\item{DE_info_emp}{Output of `get_empirical_pvals`}

\item{adj_pval}{Should the adjusted p-values be compared (TRUE) or the non-adjusted ones (FALSE)? Defautl: FALSE}
}
\value{
a list òf plots for each celltype-contrast pair: an upset plot and line plot are drawn.
}
\description{
\code{compare_normal_emp_pvals}  Compare nr and rank of DE genes between normal p-values and empirical p-values
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
batches = NA
covariates = NA
contrasts_oi = c("'High-Low','Low-High'")
senders_oi = SummarizedExperiment::colData(sce)[,celltype_id] \%>\% unique()
receivers_oi = SummarizedExperiment::colData(sce)[,celltype_id] \%>\% unique()
DE_info = get_DE_info(
   sce = sce,
   sample_id = sample_id,
   celltype_id = celltype_id,
   group_id = group_id,
   batches = batches,
   covariates = covariates,
   contrasts = contrasts_oi)
DE_info_emp = get_empirical_pvals(DE_info$celltype_de$de_output_tidy)
comparison_plots = compare_normal_emp_pvals(DE_info, DE_info_emp)
}
}
