% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{fix_frq_df}
\alias{fix_frq_df}
\title{fix_frq_df}
\usage{
fix_frq_df(sce, frq_celltype_samples)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}

\item{frq_celltype_samples}{Sample-average data frame output of `get_muscat_exprs_frac`}
}
\value{
Fixed data frame with fraction of cells expressing a gene.
}
\description{
\code{fix_frq_df}  Fix muscat-bug in fraction calculation in case that expression fraction would be NA / NaN. Change NA to 0.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
frq_df = get_muscat_exprs_frac(sce, sample_id = sample_id, celltype_id =  celltype_id, group_id = group_id) \%>\% .$frq_celltype_samples
if(nrow(frq_df \%>\% dplyr::filter(is.na(fraction_sample))) > 0 | nrow(frq_df \%>\% dplyr::filter(is.nan(fraction_sample))) > 0) {
  frq_df = fix_frq_df(sce, frq_df)
  }
}

}
