% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{generate_prioritization_tables}
\alias{generate_prioritization_tables}
\title{generate_prioritization_tables}
\usage{
generate_prioritization_tables(sender_receiver_info, sender_receiver_de, ligand_activities_targets_DEgenes, contrast_tbl, sender_receiver_tbl, grouping_tbl, prioritizing_weights, fraction_cutoff)
}
\arguments{
\item{sender_receiver_info}{XXX}

\item{sender_receiver_de}{XXX}

\item{ligand_activities_targets_DEgenes}{XXX}

\item{contrast_tbl}{XXXX}

\item{sender_receiver_tbl}{XXX}

\item{grouping_tbl}{XXX}

\item{prioritizing_weights}{XXXX}
}
\value{
XXXX
}
\description{
\code{generate_prioritization_tables}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
