% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{get_DE_info_sampleAgnostic}
\alias{get_DE_info_sampleAgnostic}
\title{get_DE_info_sampleAgnostic}
\usage{
get_DE_info_sampleAgnostic(sce, group_id, celltype_id, contrasts_oi, expressed_df, min_cells = 10, contrast_tbl)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{contrasts_oi}{String indicating the contrasts of interest (= which groups/conditions will be compared) for the differential expression and MultiNicheNet analysis. 
We will demonstrate here a few examples to indicate how to write this. Check the limma package manuals for more information about defining design matrices and contrasts for differential expression analysis. \cr
If wanting to compare group A vs B: `contrasts_oi = c("'A-B'")` \cr
If wanting to compare group A vs B & B vs A: `contrasts_oi = c("'A-B','B-A'")` \cr
If wanting to compare group A vs B & A vs C & A vs D: `contrasts_oi = c("'A-B','A-C', 'A-D'")` \cr
If wanting to compare group A vs B and C: `contrasts_oi = c("'A-(B+C)/2'")` \cr
If wanting to compare group A vs B, C and D: `contrasts_oi = c("'A-(B+C+D)/3'")` \cr
If wanting to compare group A vs B, C and D & B vs A,C,D: `contrasts_oi = c("'A-(B+C+D)/3', 'B-(A+C+D)/3'")` \cr
Note that the groups A, B, ... should be present in the meta data column 'group_id'.}

\item{expressed_df}{tibble with three columns: gene, celltype, expressed; this data frame indicates which genes can be considered as expressed in each cell type.}

\item{min_cells}{Indicates the minimal number of cells that a sample should have to be considered in the DE analysis. Default: 10. See `muscat::pbDS`.}

\item{contrast_tbl}{see explanation in multi_nichenet_analysis function -- here: only required to give as input if findMarkers = TRUE.}
}
\value{
List with output of the differential expression analysis in 1) default format(`muscat::pbDS()`), and 2) in a tidy table format (`muscat::resDS()`) (both in the `celltype_de` slot); Histogram plot of the p-values is also returned.
}
\description{
\code{get_DE_info_sampleAgnostic} Perform differential expression analysis via scran::findMarkers approach. Also visualize the p-value distribution.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
batches = NA
covariates = NA
contrasts_oi = c("'High-Low','Low-High'")
contrast_tbl = tibble(contrast = c("High-Low","Low-High"), group = c("High","Low"))
frq_list = get_frac_exprs_sampleAgnostic(sce = sce, sample_id = sample_id, celltype_id =  celltype_id, group_id = group_id)
DE_info = get_DE_info_sampleAgnostic(
   sce = sce,
   celltype_id = celltype_id,
   group_id = group_id,
   contrasts = contrasts_oi,
   expressed_df = frq_list$expressed_df,
   contrast_tbl = contrast_tbl)
}

}
