% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscat_de.R
\name{get_FDR_empirical}
\alias{get_FDR_empirical}
\title{Add empirical p-values and adjusted p-values to a subset of the DE output table.}
\usage{
get_FDR_empirical(de_output_tidy, cluster_id_oi, contrast_oi)
}
\arguments{
\item{de_output_tidy}{Data frame of DE results, containing at least the following columns: cluster_id, contrast, p_val, logFC.}

\item{cluster_id_oi}{Indicate which celltype DE results should be filtered for.}

\item{contrast_oi}{Indicate which contrast DE results should be filtered for.}
}
\value{
de_output_tidy dataframe (for the celltype and contrast of interest) with two columns added: p_emp and p_adj_emp.
}
\description{
\code{get_FDR_empirical}  Add empirical p-values and adjusted p-values to a subset of the DE output table. This is the function that works under  the hood of `add_empirical_pval_fdr`. Credits to Jeroen Gillis (cf satuRn package)
}
