% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscat_de.R
\name{get_FDR_empirical_plots_all}
\alias{get_FDR_empirical_plots_all}
\title{Get diagnostic plots of the empirical null.}
\usage{
get_FDR_empirical_plots_all(de_output_tidy)
}
\arguments{
\item{de_output_tidy}{Data frame of DE results, containing at least the following columns: cluster_id, contrast, p_val, logFC.}
}
\value{
list of plots
}
\description{
\code{get_FDR_empirical_plots_all}  Get diagnostic plots of the empirical null. Credits to Jeroen Gillis (cf satuRn package)
}
