% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{get_frac_exprs_sampleAgnostic}
\alias{get_frac_exprs_sampleAgnostic}
\title{get_frac_exprs_sampleAgnostic}
\usage{
get_frac_exprs_sampleAgnostic(sce, sample_id, celltype_id, group_id, batches = NA, min_cells = 10, fraction_cutoff = 0.05, min_sample_prop = 0.5)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}

\item{sample_id}{Name of the meta data column that indicates from which sample/patient a cell comes from}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from}

\item{batches}{NA if no batches should be corrected for. If there should be corrected for batches during DE analysis and pseudobulk expression calculation, this argument should be the name(s) of the columns in the meta data that indicate the batch(s). Should be categorical. Pseudobulk expression values will be corrected for the first element of this vector.}

\item{min_cells}{Indicates the minimal number of cells that a sample should have to be considered in the DE analysis. Default: 10. See `muscat::pbDS`.}

\item{fraction_cutoff}{Cutoff indicating the minimum fraction of cells of a cell type in a specific sample that are necessary to consider a gene (e.g. ligand/receptor) as expressed in a sample.}

\item{min_sample_prop}{Default, and only recommended value = 1. Hereby, the gene should be expressed in at least one group/condition.}
}
\value{
List containing data frames with the fraction of expression per sample and per group.
}
\description{
\code{get_frac_exprs_sampleAgnostic}  Calculate the average fraction of expression of each gene per group. All cells from all samples will be pooled per group/condition.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
frac_info = get_frac_exprs_sampleAgnostic(sce = sce, sample_id = sample_id, celltype_id =  celltype_id, group_id = group_id)
}

}
