% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_activities.R
\name{get_ligand_activities_targets_DEgenes}
\alias{get_ligand_activities_targets_DEgenes}
\title{get_ligand_activities_targets_DEgenes}
\usage{
get_ligand_activities_targets_DEgenes(receiver_de, receivers_oi, receiver_frq_df_group, ligand_target_matrix, logFC_threshold = 0.25, p_val_threshold = 0.05, frac_cutoff = 0.05, p_val_adj = FALSE, top_n_target = 250)
}
\arguments{
\item{receiver_de}{XXX}

\item{receivers_oi}{XXX}

\item{receiver_frq_df_group}{XXX}

\item{ligand_target_matrix}{XXXX}

\item{logFC_threshold}{XXXX}

\item{p_val_threshold}{XXXX}

\item{frac_cutoff}{XXXX}

\item{p_val_adj}{XXXX}

\item{top_n_target}{XXXX}
}
\value{
XXXX
}
\description{
\code{get_ligand_activities_targets_DEgenes}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
