% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{get_muscat_exprs_avg}
\alias{get_muscat_exprs_avg}
\title{get_muscat_exprs_avg}
\usage{
get_muscat_exprs_avg(seurat_obj, sample_id, celltype_id, group_id, assay_oi_sce = "RNA")
}
\arguments{
\item{seurat_obj}{XXXX}

\item{sample_id}{XXXX}

\item{celltype_id}{XXXX}

\item{group_id}{XXXX}

\item{assay_oi_sce}{XXXX}
}
\value{
XXXX
}
\description{
\code{get_muscat_exprs_avg}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
