% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{inner_get_muscat_exprs_frac_loop_1}
\alias{inner_get_muscat_exprs_frac_loop_1}
\title{inner_get_muscat_exprs_frac_loop_1}
\usage{
inner_get_muscat_exprs_frac_loop_1(celltype_oi, frq, all_genes, samples)
}
\arguments{
\item{samples}{}
}
\description{
inner_get_muscat_exprs_frac_loop_1
}
