% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligand_target_matrix_test}
\alias{ligand_target_matrix_test}
\title{Ligand-Target Matrix: subset NicheNet 2.0.}
\format{
A matrix
}
\usage{
ligand_target_matrix_test
}
\description{
Matrix with ligand-target regulatory potential scores
}
\keyword{datasets}
