% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_circos_group_comparison}
\alias{make_circos_group_comparison}
\title{make_circos_group_comparison}
\usage{
make_circos_group_comparison(prioritized_tbl_oi, colors_sender, colors_receiver)
}
\arguments{
\item{prioritized_tbl_oi}{XXX}

\item{colors_sender}{XXX}

\item{colors_receiver}{XXX}
}
\value{
XXXX
}
\description{
\code{make_circos_group_comparison}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
