% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ggraph_signaling_path}
\alias{make_ggraph_signaling_path}
\title{make_ggraph_signaling_path}
\usage{
make_ggraph_signaling_path(signaling_graph_list, colors, ligands_all, receptors_all, targets_all)
}
\arguments{
\item{signaling_graph_list}{Output of `nichenetr::get_ligand_signaling_path_with_receptor`}

\item{colors}{Named vector of colors associated to each node type: Example: colors <- c("ligand" = "indianred2", "receptor" = "orange", "target" = "steelblue2", "mediator" = "grey25").}

\item{ligands_all}{Name of the ligand(s)}

\item{receptors_all}{Name of the receptor(s)}

\item{targets_all}{Name of the target(s)}
}
\value{
ggraph and tidygraph objec of signaling paths between predefined LR-->Target links
}
\description{
\code{make_ggraph_signaling_path} Visualize the Ligand-Receptor to target signaling paths
}
