% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_group_lfc_exprs_activity_plot}
\alias{make_group_lfc_exprs_activity_plot}
\title{make_group_lfc_exprs_activity_plot}
\usage{
make_group_lfc_exprs_activity_plot(prioritization_tables, prioritized_tbl_oi, receiver_oi, heights = NULL)
}
\arguments{
\item{prioritization_tables}{XXX}

\item{prioritized_tbl_oi}{XXX}

\item{receiver_oi}{XXX}

\item{heights}{XXX}
}
\value{
XXXX
}
\description{
\code{make_group_lfc_exprs_activity_plot}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
