% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ligand_activity_plots}
\alias{make_ligand_activity_plots}
\title{make_ligand_activity_plots}
\usage{
make_ligand_activity_plots(prioritization_tables, ligands_oi, contrast_tbl, widths = NULL)
}
\arguments{
\item{prioritization_tables}{XXX}

\item{ligands_oi}{Character vector of ligands for which the activities should be visualized}

\item{contrast_tbl}{Table to link the contrast definitions to the group ids.}

\item{widths}{Vector of 2 elements: Width of the scaled ligand activity panel, width of the ligand activity panel. Default NULL: automatically defined based number of group-receiver combinations. If manual change: example format: c(3,2)}
}
\value{
Heatmap of ligand activities (normal and scaled) of each group-receiver combination
}
\description{
\code{make_ligand_activity_plots}  Visualize the ligand activities (normal and scaled) of each group-receiver combination
}
\examples{
\dontrun{
print("XXXX")
}

}
