% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ligand_activity_target_plot}
\alias{make_ligand_activity_target_plot}
\title{make_ligand_activity_target_plot}
\usage{
make_ligand_activity_target_plot(group_oi, receiver_oi, prioritized_tbl_oi, ligand_activities_targets_DEgenes, contrast_tbl, grouping_tbl, receiver_info, plot_legend = TRUE, heights = NULL, widths = NULL)
}
\arguments{
\item{group_oi}{XXX}

\item{receiver_oi}{XXX}

\item{prioritized_tbl_oi}{XXX}

\item{ligand_activities_targets_DEgenes}{XXX}

\item{contrast_tbl}{XXX}

\item{grouping_tbl}{XXX}

\item{receiver_info}{XXX}

\item{plot_legend}{XXX}

\item{heights}{XXX}

\item{widths}{XXX}
}
\value{
XXXX
}
\description{
\code{make_ligand_activity_target_plot}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
