% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ligand_receptor_nebulosa_feature_plot}
\alias{make_ligand_receptor_nebulosa_feature_plot}
\title{make_ligand_receptor_nebulosa_feature_plot}
\usage{
make_ligand_receptor_nebulosa_feature_plot(seurat_obj_sender, seurat_obj_receiver, ligand_oi, receptor_oi, group_oi, group_id, celltype_id_sender, celltype_id_receiver, senders_oi, receivers_oi, prioritized_tbl_oi)
}
\arguments{
\item{seurat_obj_sender}{XXX}

\item{seurat_obj_receiver}{XXX}

\item{ligand_oi}{XXX}

\item{receptor_oi}{XXX}

\item{group_oi}{XXX}

\item{group_id}{XXX}

\item{celltype_id_sender}{XXX}

\item{celltype_id_receiver}{XXX}

\item{senders_oi}{XXX}

\item{receivers_oi}{XXX}

\item{prioritized_tbl_oi}{XXX}
}
\value{
XXXX
}
\description{
\code{make_ligand_receptor_nebulosa_feature_plot}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
