% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{make_lite_output_condition_specific}
\alias{make_lite_output_condition_specific}
\title{make_lite_output_condition_specific}
\usage{
make_lite_output_condition_specific(multinichenet_output, top_n_LR = 2500)
}
\arguments{
\item{multinichenet_output}{Output of a MultiNicheNet analysis (result of `multi_nichenet_analysis()`).}

\item{top_n_LR}{top nr of LR pairs for which correlation with target genes will be calculated. Is 2500 by default. If you want to calculate correlation for all LR pairs, set this argument to NA.}
}
\value{
multinichenet output list (= result of `multi_nichenet_analysis()`), but now filtered such that expression information is only returned for present ligands, receptors, and genes DE in at least one probed condition.
}
\description{
\code{make_lite_output_condition_specific} Reduce the size of the MultiNicheNet output object (for memory efficiency), by only keeping expression information for present ligands, receptors, and genes DE in at least one probed condition. This function is specific for an object that contains the prioritization tables of the condition-specific cell type workflow as well.
}
\examples{
\dontrun{
library(dplyr)
multinichenet_output = multinichenet_output \%>\% make_lite_output_condition_specific()

}

}
