% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_nebulosa}
\alias{make_nebulosa}
\title{make_nebulosa}
\usage{
make_nebulosa(seurat_subset_oi, seurat_subset_bg, title_umap, gene_oi, group_oi, background_groups)
}
\arguments{
\item{seurat_subset_oi}{XXX}

\item{seurat_subset_bg}{XXX}

\item{title_umap}{XXX}

\item{gene_oi}{XXX}

\item{group_oi}{XXX}

\item{background_groups}{XXX}
}
\value{
XXXX
}
\description{
\code{make_nebulosa}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
