% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_sample_lr_prod_activity_plots}
\alias{make_sample_lr_prod_activity_plots}
\title{make_sample_lr_prod_activity_plots}
\usage{
make_sample_lr_prod_activity_plots(prioritization_tables, prioritized_tbl_oi, widths = NULL)
}
\arguments{
\item{prioritization_tables}{XXX}

\item{prioritized_tbl_oi}{XXX}

\item{widths}{Vector of 3 elements: Width of the LR exprs product panel,  width of the scaled ligand activity panel, width of the ligand activity panel. Default NULL: automatically defined based on nr of samples vs nr of group-receiver combinations. If manual change: example format: c(5,1,1)}
}
\value{
Ligand-Receptor Expression Product & Ligand Activities Dotplot/Heatmap
}
\description{
\code{make_sample_lr_prod_activity_plots}  Visualize the scaled product of Ligand-Receptor (pseudobulk) expression per sample, and compare the different groups. In addition, show the NicheNet ligand activities in each receiver-celltype combination.
}
\examples{
\dontrun{
print("XXXX")
}

}
