% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_sample_target_plots}
\alias{make_sample_target_plots}
\title{make_sample_target_plots}
\usage{
make_sample_target_plots(receiver_info, targets_oi, receiver_oi, grouping_tbl)
}
\arguments{
\item{receiver_info}{XXX}

\item{targets_oi}{Character vector of genes of which the expression should be visualized}

\item{receiver_oi}{Name of the receiver cell type of interest for which expression of genes should be visualized}

\item{grouping_tbl}{Data frame linking the sample_id, group_id and covariate_oi}
}
\value{
Heatmap/Dotplot of scaled target gene expression (samples in columns, genes in rows)
}
\description{
\code{make_sample_target_plots}  Heatmap/Dotplot of scaled target gene expression per sample, compared between the groups of interest. (samples in columns, genes in rows)
}
\examples{
\dontrun{
print("XXXX")
}

}
