% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_target_feature_plot}
\alias{make_target_feature_plot}
\title{make_target_feature_plot}
\usage{
make_target_feature_plot(sce_receiver, target_oi, group_oi, group_id, celltype_id_receiver, receivers_oi, prioritized_tbl_oi)
}
\arguments{
\item{sce_receiver}{SingleCellExperiment object containing the receiver cell types of interest}

\item{target_oi}{Character vector of the name of the target gene of interest}

\item{group_oi}{Character vector of name of the group of interest}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from (in both sce_receiver and sce_sender)}

\item{celltype_id_receiver}{Name of the meta data column that indicates the cell type of a cell (in sce_receiver).}

\item{receivers_oi}{Character vector with the names of the receiver cell types of interest}

\item{background_groups}{Default NULL: all groups in the group_id metadata column will be chosen. But user can fill in a character vector with the names of all gruops of interest.}
}
\value{
UMAP showing target gene expression in groups and celltypes of interest vs other groups and cell types.
}
\description{
\code{make_target_feature_plot}  UMAP showing target gene expression in groups and celltypes of interest vs other groups and cell types.
}
\examples{
\dontrun{
print("XXXX")
}

}
