% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_target_nebulosa_feature_plot}
\alias{make_target_nebulosa_feature_plot}
\title{make_target_nebulosa_feature_plot}
\usage{
make_target_nebulosa_feature_plot(seurat_obj_receiver, target_oi, group_oi, group_id, celltype_id_receiver, receivers_oi, prioritized_tbl_oi)
}
\arguments{
\item{seurat_obj_receiver}{XXX}

\item{target_oi}{XXX}

\item{group_oi}{XXX}

\item{group_id}{XXX}

\item{celltype_id_receiver}{XXX}

\item{receivers_oi}{XXX}

\item{prioritized_tbl_oi}{XXX}
}
\value{
XXXX
}
\description{
\code{make_target_nebulosa_feature_plot}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
