% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{makenames_SCE}
\alias{makenames_SCE}
\title{make.names of all genes in a SingleCellExperiment Object}
\usage{
makenames_SCE(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}
}
\value{
SingleCellExperiment Object
}
\description{
\code{makenames_SCE} make.names of all genes in a SingleCellExperiment Object. Avoids missing of genes that are sometimes in original symbol, and sometimes in make.names() format
}
\examples{
\dontrun{
sce = sce \%>\% makenames_SCE()
}

}
