% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{ms_mg_nichenet_analysis}
\alias{ms_mg_nichenet_analysis}
\title{ms_mg_nichenet_analysis}
\usage{
ms_mg_nichenet_analysis(sender_receiver_separate = TRUE, ...)
}
\arguments{
\item{sender_receiver_separate}{XXXX}

\item{...}{Arguments to `ms_mg_nichenet_analysis_separate` (default; when `sender_receiver_separate = TRUE`) or `ms_mg_nichenet_analysis_combined` (when `sender_receiver_separate = FALSE`)}
}
\value{
XXXX
}
\description{
\code{ms_mg_nichenet_analysis}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
