% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{ms_mg_nichenet_analysis_separate}
\alias{ms_mg_nichenet_analysis_separate}
\title{ms_mg_nichenet_analysis_separate}
\usage{
ms_mg_nichenet_analysis_separate(
seurat_obj_receiver,seurat_obj_sender,celltype_id_receiver,celltype_id_sender,sample_id,group_id,lr_network,ligand_target_matrix,contrasts_oi,contrast_tbl,
prioritizing_weights = c("scaled_lfc_ligand" = 1, "scaled_p_val_ligand" = 1, "scaled_lfc_receptor" = 1, "scaled_p_val_receptor" = 1, "scaled_activity_scaled" = 1.5,
"scaled_activity" = 0.5,"scaled_avg_exprs_ligand" = 1,"scaled_avg_frq_ligand" = 1,"scaled_avg_exprs_receptor" = 1, "scaled_avg_frq_receptor" = 1,
"fraction_expressing_ligand_receptor" = 1,"scaled_abundance_sender" = 0, "scaled_abundance_receiver" = 0),
assay_oi_sce = "RNA",assay_oi_pb ="counts",fun_oi_pb = "sum",de_method_oi = "edgeR",min_cells = 10,logFC_threshold = 0.25,p_val_threshold = 0.05,frac_cutoff = 0.05,p_val_adj = FALSE,top_n_target = 250, verbose = TRUE)
}
\arguments{
\item{seurat_obj_receiver}{XXXX}

\item{seurat_obj_sender}{XXXX}

\item{celltype_id_receiver}{XXXX}

\item{celltype_id_sender}{XXXX}

\item{sample_id}{XXXX}

\item{group_id}{XXXX}

\item{lr_network}{XXXX}

\item{ligand_target_matrix}{XXXX}

\item{contrasts_oi}{XXXX}

\item{contrast_tbl}{XXXX}

\item{prioritizing_weights}{XXXX}

\item{assay_oi_sce}{XXXX}

\item{assay_oi_pb}{XXXX}

\item{fun_oi_pb}{XXXX}

\item{de_method_oi}{XXXX}

\item{min_cells}{XXXX}

\item{logFC_threshold}{XXXX}

\item{p_val_threshold}{XXXX}

\item{frac_cutoff}{XXXX}

\item{p_val_adj}{XXXX}

\item{top_n_target}{XXXX}

\item{verbose}{XXXX}
}
\value{
XXXX
}
\description{
\code{ms_mg_nichenet_analysis_separate}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
