% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_analysis.R
\name{pca_LR_pb_prod_matrix}
\alias{pca_LR_pb_prod_matrix}
\title{pca_LR_pb_prod_matrix}
\usage{
pca_LR_pb_prod_matrix(lr_prod_mat, metadata_tbl)
}
\arguments{
\item{lr_prod_mat}{Output of `calculate_LR_pb_prod_matrix`}

\item{metadata_tbl}{metadata tibble, with at least the column "sample_id"}
}
\value{
list with following elements: \cr
`pca_lr_prod_mat`: PCA output \cr
`LR_pair_contribution_tbl`: loadings of each ligand-receptor-sender-receiver pair in contributing to each principal component (PC) \cr
`coordinate_tbl`: coordinates of each PC \cr
`metadata_correlation_tbl`: correlation of each PC with each metadata column of interest (in metadata_tbl) \cr
}
\description{
\code{pca_LR_pb_prod_matrix} Perform PCA and downstream analysis to analyze sample-to-sample heterogeneity in ligand-receptor pair expression.
}
\examples{
\dontrun{
library(dplyr)
lr_network = readRDS(url("https://zenodo.org/record/3260758/files/lr_network.rds"))
lr_network = lr_network \%>\% dplyr::rename(ligand = from, receptor = to) \%>\% dplyr::distinct(ligand, receptor)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
senders_oi = "CAF"
receivers_oi = "Malignant"
fraction_cutoff = 0.05
batches = NA
lr_prod_mat = calculate_LR_pb_prod_matrix(sce, sample_id, celltype_id, group_id, senders_oi, receivers_oi, fraction_cutoff, lr_network, batches)
metadata_tbl = SummarizedExperiment::colData(sce)[, c(sample_id, group_id)]  \%>\% data.frame() \%>\% tibble::as_tibble() \%>\% distinct()
colnames(metadata_tbl) = c("sample_id","group_id")
metadata_tbl2 = metadata_tbl \%>\% mutate(group_id = as.double(factor(group_id)))
output_pca = pca_LR_pb_prod_matrix(lr_prod_mat, metadata_tbl)
}

}
