% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscat_de.R
\name{perform_muscat_de_analysis}
\alias{perform_muscat_de_analysis}
\title{perform_muscat_de_analysis}
\usage{
perform_muscat_de_analysis(seurat_obj, sample_id, celltype_id, group_id, covariates, contrasts, assay_oi_sce = "RNA", assay_oi_pb = "counts", fun_oi_pb = "sum", de_method_oi = "edgeR", min_cells = 10)
}
\arguments{
\item{seurat_obj}{XXXX}

\item{sample_id}{XXXX}

\item{celltype_id}{XXXX}

\item{group_id}{XXXX}

\item{assay_oi_sce}{XXXX}

\item{assay_oi_pb}{XXXX}

\item{fun_oi_pb}{XXXX}

\item{de_method_oi}{XXXX}

\item{min_cells}{XXXX}
}
\value{
XXXX
}
\description{
\code{perform_muscat_de_analysis}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
