% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_obj}
\alias{seurat_obj}
\title{Seurat object containing scRNAseq data (subsampled)}
\format{
An object of class Seurat
}
\usage{
seurat_obj
}
\description{
Seurat object containing scRNAseq data (subsampled). Source of the data: Puram et al., Cell 2017: “Single-Cell Transcriptomic Analysis of Primary and Metastatic Tumor Ecosystems in Head and Neck Cancer.”. This example data was downsampled (features and cells).
}
\keyword{datasets}
