% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{fix_frq_df}
\alias{fix_frq_df}
\title{fix_frq_df}
\usage{
fix_frq_df(seurat_obj, frq_celltype_samples)
}
\arguments{
\item{seurat_obj}{XXXX}

\item{frq_celltype_samples}{XXXX}
}
\value{
XXXX
}
\description{
\code{fix_frq_df}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
