% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{get_avg_frac_exprs_abund}
\alias{get_avg_frac_exprs_abund}
\title{get_avg_frac_exprs_abund}
\usage{
get_avg_frac_exprs_abund(seurat_obj, sample_id, celltype_id, group_id, assay_oi = "RNA")
}
\arguments{
\item{seurat_obj}{XXXX}

\item{sample_id}{XXXX}

\item{celltype_id}{XXXX}

\item{group_id}{XXXX}
}
\value{
XXXX
}
\description{
\code{get_avg_frac_exprs_abund}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
