% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{get_muscat_exprs_frac}
\alias{get_muscat_exprs_frac}
\title{get_muscat_exprs_frac}
\usage{
get_muscat_exprs_frac(sce, sample_id, celltype_id, group_id)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest. Contains both sender and receiver cell types.}

\item{sample_id}{Name of the meta data column that indicates from which sample/patient a cell comes from}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from}
}
\value{
List with two dataframes: one with fraction of cells in a cell type expressing a gene, averaged per sample; and one averaged per group.
}
\description{
\code{get_muscat_exprs_frac} Calculate sample- and group-average of fraction of cells in a cell type expressing a gene.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
muscat_exprs_frac = get_muscat_exprs_frac(sce = sce, sample_id = sample_id, celltype_id =  celltype_id, group_id = group_id)
}

}
