% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_featureplot}
\alias{make_featureplot}
\title{make_featureplot}
\usage{
make_featureplot(seurat_subset_oi, title_umap, gene_oi, group_oi, background_groups, group_id)
}
\arguments{
\item{seurat_subset_oi}{XXX}

\item{title_umap}{XXX}

\item{gene_oi}{XXX}

\item{group_oi}{XXX}

\item{background_groups}{XXX}

\item{group_id}{}
}
\value{
XXXX
}
\description{
\code{make_featureplot}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
