% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_group_lfc_exprs_activity_plot}
\alias{make_group_lfc_exprs_activity_plot}
\title{make_group_lfc_exprs_activity_plot}
\usage{
make_group_lfc_exprs_activity_plot(prioritization_tables, prioritized_tbl_oi, receiver_oi, heights = NULL)
}
\arguments{
\item{prioritization_tables}{Output of `generate_prioritization_tables` or sublist in the output of `multi_nichenet_analysis`}

\item{prioritized_tbl_oi}{Subset of `prioritization_tables$group_prioritization_tbl`: the ligand-receptor interactions shown in this subset will be visualized: recommended to consider the top n LR interactions of a group of interest, based on the prioritization_score (eg n = 50; see vignettes for examples).}

\item{receiver_oi}{Name of the receiver cell type of interest for which ligand-receptor pair LFC and ligand activities should be shown}

\item{heights}{Vector of 3 elements: Height of the LFC panel, normal ligand activity, and scaled ligand activity panel. Default NULL: automatically defined based on the nr of groups and sender-receiver combinations. If manual change: example format: c(5,1,1)}
}
\value{
Summary plot showing the logFC of each ligand-receptor pair / group combination, complemented by the NicheNet scaled and normal ligand activities
}
\description{
\code{make_group_lfc_exprs_activity_plot}  Summary plot showing the logFC of each ligand-receptor pair / group combination, complemented by the NicheNet scaled and normal ligand activities
}
\examples{
\dontrun{
library(dplyr)
lr_network = readRDS(url("https://zenodo.org/record/3260758/files/lr_network.rds"))
lr_network = lr_network \%>\% dplyr::rename(ligand = from, receptor = to) \%>\% dplyr::distinct(ligand, receptor)
ligand_target_matrix = readRDS(url("https://zenodo.org/record/3260758/files/ligand_target_matrix.rds"))
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
covariates = NA
contrasts_oi = c("'High-Low','Low-High'")
contrast_tbl = tibble(contrast = c("High-Low","Low-High"), group = c("High","Low"))
output = multi_nichenet_analysis(
     sce = sce, 
     celltype_id = celltype_id, 
     sample_id = sample_id, 
     group_id = group_id,
     covariates = covariates,
     lr_network = lr_network, 
     ligand_target_matrix = ligand_target_matrix, 
     contrasts_oi = contrasts_oi, 
     contrast_tbl = contrast_tbl, 
     sender_receiver_separate = FALSE
     )
group_oi = "High"
receiver_oi = "Malignant"
prioritized_tbl_oi = output$prioritization_tables$group_prioritization_tbl \%>\% filter(fraction_expressing_ligand_receptor > 0) \%>\% filter(group == group_oi) \%>\% top_n(50, prioritization_score) 
plot_oi = make_group_lfc_exprs_activity_plot(output$prioritization_tables, prioritized_tbl_oi, receiver_oi = receiver_oi)
plot_oi
}

}
