% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ligand_activity_plots}
\alias{make_ligand_activity_plots}
\title{make_ligand_activity_plots}
\usage{
make_ligand_activity_plots(prioritization_tables, ligands_oi, contrast_tbl, widths = NULL)
}
\arguments{
\item{prioritization_tables}{XXX}

\item{ligands_oi}{XXX}

\item{contrast_tbl}{XXX}

\item{widths}{XXX}
}
\value{
XXXX
}
\description{
\code{make_ligand_activity_plots}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
