% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{make_lite_output}
\alias{make_lite_output}
\title{make_lite_output}
\usage{
make_lite_output(multinichenet_output)
}
\arguments{
\item{multinichenet_output}{Output of a MultiNicheNet analysis (result of `multi_nichenet_analysis()`).}
}
\value{
multinichenet output list (= result of `multi_nichenet_analysis()`), but now filtered such that expression information is only returned for present ligands, receptors, and genes DE in at least one probed condition.
}
\description{
\code{make_lite_output} Reduce the size of the MultiNicheNet output object (for memory efficiency), by only keeping expression information for present ligands, receptors, and genes DE in at least one probed condition.
}
\examples{
\dontrun{
library(dplyr)
multinichenet_output = multinichenet_output \%>\% make_lite_output()

}

}
