% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_sample_lr_prod_activity_covariate_plots}
\alias{make_sample_lr_prod_activity_covariate_plots}
\title{make_sample_lr_prod_activity_covariate_plots}
\usage{
make_sample_lr_prod_activity_covariate_plots(prioritization_tables, prioritized_tbl_oi, grouping_tbl, covariate_oi, widths = NULL, heights = NULL)
}
\arguments{
\item{prioritization_tables}{XXX}

\item{prioritized_tbl_oi}{XXX}

\item{widths}{XXX}
}
\value{
XXXX
}
\description{
\code{make_sample_lr_prod_activity_covariate_plots}  XXXX EXPERIMENTAL FEATURE!!!!!!!!!!!!!!!!!!!!NOT WORKINGN RIGHT NOW
}
\examples{
\dontrun{
print("XXXX")
}

}
