% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_sample_lr_prod_plots}
\alias{make_sample_lr_prod_plots}
\title{make_sample_lr_prod_plots}
\usage{
make_sample_lr_prod_plots(prioritization_tables, prioritized_tbl_oi)
}
\arguments{
\item{prioritization_tables}{XXX}

\item{prioritized_tbl_oi}{XXX}
}
\value{
Ligand-Receptor Expression Product Dotplot/Heatmap
}
\description{
\code{make_sample_lr_prod_plots}  Visualize the scaled product of Ligand-Receptor (pseudobulk) expression per sample, and compare the different groups
}
\examples{
\dontrun{
print("XXXX")
}

}
