% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_sample_target_plots}
\alias{make_sample_target_plots}
\title{make_sample_target_plots}
\usage{
make_sample_target_plots(receiver_info, targets_oi, receiver_oi, grouping_tbl)
}
\arguments{
\item{receiver_info}{XXX}

\item{targets_oi}{XXX}

\item{receiver_oi}{XXX}

\item{grouping_tbl}{XXX}
}
\value{
XXXX
}
\description{
\code{make_sample_target_plots}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
