% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_target_violin_plot}
\alias{make_target_violin_plot}
\title{make_target_violin_plot}
\usage{
make_target_violin_plot(sce_receiver, target_oi, receiver_oi, group_oi, group_id, sample_id, celltype_id_receiver, covariate_oi = NA, background_groups = NULL)
}
\arguments{
\item{sce_receiver}{SingleCellExperiment object containing the receiver cell types of interest}

\item{target_oi}{Name of the gene of interest}

\item{receiver_oi}{Character vector with the names of the receiver cell type of interest}

\item{group_oi}{Character vector of name of the group of interest}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from (in both sce_receiver and sce_sender)}

\item{sample_id}{Name of the colData(sce) column in which the id of the sample is defined}

\item{celltype_id_receiver}{Name of the meta data column that indicates the cell type of a cell (in sce_receiver).}

\item{covariate_oi}{Name of a covariate of interest based on which the visualization will be split. Default: NA: no covariate.}

\item{background_groups}{Default NULL: all groups in the group_id metadata column will be chosen. But user can fill in a character vector with the names of all gruops of interest.}
}
\value{
ggplot object: Violin plot of a target gene of interest: per sample, and samples are grouped per group
}
\description{
\code{make_target_violin_plot}  Violin plot of a target gene of interest: per sample, and samples are grouped per group
}
\examples{
\dontrun{
library(dplyr)
lr_network = readRDS(url("https://zenodo.org/record/3260758/files/lr_network.rds"))
lr_network = lr_network \%>\% dplyr::rename(ligand = from, receptor = to) \%>\% dplyr::distinct(ligand, receptor)
ligand_target_matrix = readRDS(url("https://zenodo.org/record/3260758/files/ligand_target_matrix.rds"))
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
covariates = NA
contrasts_oi = c("'High-Low','Low-High'")
contrast_tbl = tibble(contrast = c("High-Low","Low-High"), group = c("High","Low"))
output = multi_nichenet_analysis(
     sce = sce, 
     celltype_id = celltype_id, 
     sample_id = sample_id, 
     group_id = group_id,
     covariates = covariates,
     lr_network = lr_network, 
     ligand_target_matrix = ligand_target_matrix, 
     contrasts_oi = contrasts_oi, 
     contrast_tbl = contrast_tbl, 
     sender_receiver_separate = FALSE
     )
receiver_oi = "Malignant"
group_oi = "High"
target_oi = "RAB31"
p_violin_target = make_target_violin_plot(sce_receiver = sce, target_oi = target_oi, receiver_oi = receiver_oi, group_oi = group_oi, group_id = group_id, sample_id, celltype_id_receiver = celltype_id)
}
}
