% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_processing.R
\name{process_info_to_ic}
\alias{process_info_to_ic}
\title{process_info_to_ic}
\usage{
process_info_to_ic(info_object, ic_type = "sender", lr_network)
}
\arguments{
\item{info_object}{XXX}

\item{ic_type}{XXX}

\item{lr_network}{XXXX}
}
\value{
XXXX
}
\description{
\code{process_info_to_ic}  XXXX
}
\examples{
\dontrun{
print("XXXX")
}

}
